fix: circular deps in BaseCalProvider
