+++ packages/platform/atoms/cal-provider/languages.ts
@@ -27,3 +27,35 @@ export const ES = "es";
 export const NL = "nl";
 export const CAL_PROVIDER_LANGUAUES = [FR, EN, PT_BR, DE, ES, NL] as const;
 export type CalProviderLanguagesType = (typeof CAL_PROVIDER_LANGUAUES)[number];
+
+type i18nFrProps = {
+  labels?: Partial<Record<frTranslationKeys, string>>;
+  language?: "fr";
+};
+
+type i18nEnProps = {
+  labels?: Partial<Record<enTranslationKeys, string>>;
+  language?: "en";
+};
+
+type i18nPtBrProps = {
+  labels?: Partial<Record<ptBrTranslationKeys, string>>;
+  language?: "pt-BR";
+};
+
+type i18nDeProps = {
+  labels?: Partial<Record<deTranslationKeys, string>>;
+  language?: "de";
+};
+
+type i18nEsProps = {
+  labels?: Partial<Record<esTranslationKeys, string>>;
+  language?: "es";
+};
+
+type i18nNlProps = {
+  labels?: Partial<Record<nlTranslationKeys, string>>;
+  language?: "nl";
+};
+
+export type i18nProps = i18nFrProps | i18nEnProps | i18nPtBrProps | i18nDeProps | i18nEsProps | i18nNlProps;
